requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
Django==5.1.6
django-ratelimit==4.1.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
firebase-admin==6.7.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
img2pdf==0.6.0
inflection==0.5.1
lxml==5.3.2
msgpack==1.1.0
multidict==6.4.3
packaging==24.2
pikepdf==9.7.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
stripe==11.6.0
twilio==9.5.2
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.20.0
